pyproject.toml
[tool.poetry]
name = "data_sheets_schema"
version = "0.1.0"
description = "A LinkML schema for Datasheets for Datasets."
authors = ["My Name <MJoachimiak@lbl.gov>"]
license = "MIT"
readme = "README.md"
include = ["README.md", "src/data_sheets_schema/schema", "project"]

[tool.poetry.dependencies]
python = "^3.9"
linkml-runtime = "^1.1.24"
pyyaml = "^6.0.2"
jinja2 = "^3.1.6"
requests = "^2.31.0"
pydantic-ai = "^0.2.0"
beautifulsoup4 = "^4.12.0"
pdfminer-six = "^20221105"
pandas = "^2.3.1"
synapseclient = {version = "^4.9.0", python = ">=3.9,<3.14"}
jsonschema = "^4.25.1"
click = "^8.1.0"

[tool.poetry.scripts]
d4d = "data_sheets_schema.cli:cli"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[tool.poetry.group.dev.dependencies]
linkml = "^1.3.5"
mkdocs-material = "^8.2.8"
mkdocs-mermaid2-plugin = "^0.6.0"
schemasheets = "^0.4.0"
schema-automator="0.5.4rc1" 
linkml-map = "^0.3.8"
linkml-term-validator = {version = ">=0.1.1", python = ">=3.10,<4.0"}
linkml-reference-validator = {version = ">=0.1.1", python = ">=3.10,<4.0"}

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.extras]
docs = ["linkml", "mkdocs-material"]
